app/agents/meta_agents/database_agent.py
from typing import Any, Dict, List, Optional
from uuid import UUID
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
from app.models.agent.action import Action
from app.models.agent.agent import Agent
from app.db.arango_db_client import ArangoDBClient
import asyncio
import os
from dotenv import load_dotenv

load_dotenv()

class DatabaseAgent(MetaAgent):
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "database"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your database ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)

        # Initialize ArangoDB client
        self.db_client = ArangoDBClient(
            host=os.getenv("ARANGO_URL"),
            username=os.getenv("ARANGO_USERNAME"),
            password=os.getenv("ARANGO_PASSWORD")
        )
        self.db_client.connect(os.getenv("ARANGO_DB_NAME"))
        self.db_client.create_collections()

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Efficient database management is crucial for MAS performance")
        self.add_belief("Data integrity and security are top priorities")

    def _init_desires(self):
        self.add_desire("Maintain optimal database performance", priority=10)
        self.add_desire("Ensure data consistency across the MAS", priority=9)

    def _init_goals(self):
        self.add_goal("Optimize database operations", priority=9)
        self.add_goal("Implement robust data security measures", priority=8)
        self.add_goal("Ensure seamless data flow within MAS", priority=8)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Analyze current database performance",
                "Identify optimization opportunities",
                "Implement database optimizations",
                "Monitor and adjust optimizations"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for database management")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New database concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New database relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        performance_query = "What are the current database performance bottlenecks?"
        performance_issues = await self.ontology_reasoner.answer_query(performance_query)
        self.add_belief(f"Current database performance issues: {performance_issues}")

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for database management completed")

    async def plan(self):
        print("Starting planning process for database management")
        for goal in self.goals:
            if goal.description == "Optimize database operations":
                self.create_plan(
                    goal.id,
                    [
                        "Analyze query performance",
                        "Optimize database indexes",
                        "Implement query caching",
                        "Tune database configuration"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for database management completed")

    async def execute(self):
        print("Starting execution process for database management")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Database task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing database task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for database management completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(1)  # Adjust the sleep time as needed

    # Database-specific methods
    async def create_action(self, action: Action) -> Action:
        action_dict = action.dict()
        result = await self.db_client.db.collection('actions').insert(action_dict)
        action_dict['_key'] = result['_key']
        return Action(**action_dict)

    async def get_all_actions(self) -> List[Action]:
        query = "FOR a IN actions RETURN a"
        cursor = await self.db_client.db.aql.execute(query)
        return [Action(**doc) for doc in await cursor.all()]

    async def get_action(self, action_id: str) -> Optional[Action]:
        result = await self.db_client.db.collection('actions').get(action_id)
        return Action(**result) if result else None

    async def update_action(self, action: Action) -> Action:
        action_dict = action.dict()
        await self.db_client.db.collection('actions').update(str(action.id), action_dict)
        return action

    async def delete_action(self, action_id: str) -> bool:
        try:
            await self.db_client.db.collection('actions').delete(action_id)
            return True
        except:
            return False

    async def create_agent(self, agent: Agent) -> Agent:
        agent_dict = agent.dict()
        result = await self.db_client.db.collection('agents').insert(agent_dict)
        agent_dict['_key'] = result['_key']
        return Agent(**agent_dict)

    async def get_all_agents(self) -> List[Agent]:
        query = "FOR a IN agents RETURN a"
        cursor = await self.db_client.db.aql.execute(query)
        return [Agent(**doc) for doc in await cursor.all()]

    async def get_agent(self, agent_id: str) -> Optional[Agent]:
        result = await self.db_client.db.collection('agents').get(agent_id)
        return Agent(**result) if result else None

    async def update_agent(self, agent: Agent) -> Agent:
        agent_dict = agent.dict()
        await self.db_client.db.collection('agents').update(str(agent.id), agent_dict)
        return agent

    async def delete_agent(self, agent_id: str) -> bool:
        try:
            await self.db_client.db.collection('agents').delete(agent_id)
            return True
        except:
            return False

    # Additional helper methods
    async def _ensure_collections(self):
        collections = ['actions', 'agents']
        for collection in collections:
            if not await self.db_client.db.has_collection(collection):
                await self.db_client.db.create_collection(collection)

    async def initialize_database(self):
        await self._ensure_collections()
        # Add any other initialization logic here

    async def execute_query(self, query: str, bind_vars: Dict[str, Any] = None) -> List[Dict[str, Any]]:
        cursor = await self.db_client.db.aql.execute(query, bind_vars=bind_vars)
        return [doc async for doc in cursor]

    # You can add more database-specific methods here as needed